A naive mirror utility to grab the data from apis.guru for integration testing
This is the small utility that I've made with Claude Code today during the impl day,
so adding it for the posterity.

I have tried to be nice when downloading - max 3 outstanding downloads and
half a second sleep between the subsequent downloads should make it a fairly gentle
load on the apis.guru site, considering it does not re-download the existing data.

Change-Id: I420b5dc02cd72ef2ad9bb516ea49424173e632d9